requirements.txt
python-nmap
flask
waitress
bs4
requests
